handled possible issues with deleting and extracting fines

+++ app.py
@@ -119,7 +119,7 @@ def displayEmployee():
 @app.route('/dashboard/fine', methods=['GET', 'POST', 'POST_DELETE', 'GET_EXTRACT'])
 def renderFine():
     fines = runQuery("SELECT * FROM fine")
-
+    matter = runQuery("SELECT matter FROM fine")
     if request.method == 'POST':
         fine_id = runQuery("SELECT MAX(fine_id) FROM fine")[0][0]+1
         police_id = request.form['police_id']
@@ -128,11 +128,18 @@ def renderFine():
         fined_date = request.form['fined_date']
 
         runQuery("INSERT INTO fine(fine_id, police_id, aadhar_id, matter, fined_date) VALUES('{}', '{}', '{}', '{}', '{}')".format(fine_id, police_id, aadhar_id, matter, fined_date))
-        return render_template('fine_display.html', fines = fines)
+        return render_template('fine_display.html', fines = fines, matter = matter)
     
     if request.method == 'POST_DELETE':
         fine_id = request.form['fine_id']
-        runQuery("DELETE FROM fine WHERE fine_id={}".format(fine_id))
+        present_fid = runQuery("Select fine_id from fine")
+        fid_lst = ()
+        for i in range(0, len(present_fid)):
+            fid_lst.append(present_fid[i][0])
+        if fine_id in fid_lst:
+            runQuery("DELETE FROM fine WHERE fine_id={}".format(fine_id))
+        else:
+            return render_template('fine_display.html', fines = fines, message="Fine id not found")
         return render_template('fine_display.html', fines = fines)
     
     if request.method == 'GET':
@@ -140,8 +147,15 @@ def renderFine():
     
     if request.method == 'GET_EXTRACT':
         aadhar_id = request.form['aadhar_id']
-        extracted_fines = runQuery("SELECT * FROM fine WHERE aadhar_id='{}'".format(aadhar_id))
-        return render_template('fine_display.html', fines = extracted_fines)
+        present_aid = runQuery("SELECT aadhar_id FROM fine")
+        aid_lst = ()
+        for i in range(0, len(present_aid)):
+            aid_lst.append(present_aid[i][0])
+        if aadhar_id in aid_lst:
+            extracted_fines = runQuery("SELECT * FROM fine WHERE aadhar_id='{}'".format(aadhar_id))
+            return render_template('fine_display.html', fines = extracted_fines)
+        else:
+            return render_template('fine_display.html', message = "Aadhar id not found")
        
     return render_template('fine_display.html')
